+++ tests/testthat/test-RTI.R
@@ -1,10 +1,16 @@
 # Tests for RTI.R (other than rti_calc_simple())
 test_that("simple_rma() provides rma objects", {
-  testthat::expect_true("rma" %in% class(simple_rma(15, 4.74^2)))
-  testthat::expect_true("rma" %in% class(simple_rma(c(47.5, 32.5), 4.74^2)))
-  testthat::expect_true("rma" %in% class(simple_rma(jt_example_data_1, 
-                                          error_var = 4.74^2, 
-                                          observed = "obs")))
+  out1 <- NULL
+  out2 <- NULL
+  out3 <- NULL
+  expect_warning(out1 <- simple_rma(15, 4.74^2), "deprecated")
+  expect_warning(out2 <- simple_rma(c(47.5, 32.5), 4.74^2), "deprecated")
+  expect_warning(out3 <- simple_rma(jt_example_data_1,
+                                    error_var = 4.74^2,
+                                    observed = "obs"), "deprecated")
+  testthat::expect_true("rma" %in% class(out1))
+  testthat::expect_true("rma" %in% class(out2))
+  testthat::expect_true("rma" %in% class(out3))
 })
 
 # tests for compute_rti_data